requirements.txt
altgraph==0.17.4
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
cowpy==1.1.5
docker==6.1.3
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.36
pyfiglet==1.0.2
Pygments==2.17.2
pyinstaller-hooks-contrib==2023.10
python-dateutil==2.8.2
questionary==2.0.1
requests==2.31.0
rich==13.7.0
setuptools==69.0.2
six==1.16.0
termcolor==2.4.0
terminaltables==3.1.10
urllib3==2.1.0
wcwidth==0.2.12
websocket-client==1.6.4
zstandard==0.22.0
